refactor : LottoBoughtTicket 제거
상속을 함으로써 얻는 이익은 보이지 않고 로직만 복잡하게 만들기 때문에 LottoBoughtTicket 클래스는 제거

+++ src/main/java/lotto/domain/LottoResultStatistics.java
@@ -35,7 +35,7 @@ public class LottoResultStatistics {
 
     public long calculateEarning(final Money money) {
         long totalReward = Arrays.stream(LottoRank.values())
-                .mapToInt(lottoRank -> this.lottoResult.get(lottoRank) * lottoRank.getReward())
+                .mapToLong(lottoRank -> this.lottoResult.get(lottoRank) * lottoRank.getReward())
                 .sum();
 
         double paidMoney = money.getMoney();
